fetch data in hook instead, make usage components more declarative
