+++ airbyte-webapp/src/packages/cloud/views/credits/CreditsPage/components/CreditsUsage.module.scss
@@ -1,10 +1,5 @@
 @use "scss/variables";
 
-.empty {
-  height: 100%;
-  padding-bottom: variables.$spacing-xl;
-}
-
 .cardBlock {
   margin: variables.$spacing-md 0 variables.$spacing-xl;
 }
