+++ airbyte-webapp/src/packages/cloud/views/credits/CreditsPage/components/CreditsUsage.tsx
@@ -2,20 +2,15 @@ import React, { useMemo } from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 
 import { Card } from "components/ui/Card";
-import { FlexContainer } from "components/ui/Flex";
-
-import { useCurrentWorkspace } from "hooks/services/useWorkspace";
-import { useGetCloudWorkspaceUsage } from "packages/cloud/services/workspaces/CloudWorkspacesService";
 
 import styles from "./CreditsUsage.module.scss";
 import UsagePerConnectionTable from "./UsagePerConnectionTable";
 import { UsagePerDayGraph } from "./UsagePerDayGraph";
+import { useCreditsUsage } from "./useCreditsUsage";
 
 const CreditsUsage: React.FC = () => {
   const { formatDate } = useIntl();
-
-  const { workspaceId } = useCurrentWorkspace();
-  const data = useGetCloudWorkspaceUsage(workspaceId);
+  const { data } = useCreditsUsage();
 
   const chartData = useMemo(
     () =>
@@ -36,13 +31,7 @@ const CreditsUsage: React.FC = () => {
       </Card>
 
       <Card title={<FormattedMessage id="credits.usagePerConnection" />} lightPadding className={styles.cardBlock}>
-        {data?.creditConsumptionByConnector?.length ? (
-          <UsagePerConnectionTable creditConsumption={data.creditConsumptionByConnector} />
-        ) : (
-          <FlexContainer alignItems="center" justifyContent="center" className={styles.empty}>
-            <FormattedMessage id="credits.noData" />
-          </FlexContainer>
-        )}
+        <UsagePerConnectionTable />
       </Card>
     </>
   );
