+++ airbyte-webapp/src/packages/cloud/views/credits/CreditsPage/components/useCreditsUsage.tsx
@@ -1,0 +1,9 @@
+import { useGetCloudWorkspaceUsage } from "packages/cloud/services/workspaces/CloudWorkspacesService";
+import { useCurrentWorkspace } from "services/workspaces/WorkspacesService";
+
+export const useCreditsUsage = () => {
+  const { workspaceId } = useCurrentWorkspace();
+  const data = useGetCloudWorkspaceUsage(workspaceId);
+
+  return { data };
+};
